+++ src/Command/RotateCommand.php
@@ -67,7 +67,7 @@ class RotateCommand extends Command
             }
 
             foreach ($files as $file) {
-                $io->verbose(__d('database_backup', 'Backup `{0}` has been deleted', $file->get('filename')));
+                $io->verbose(__d('database_backup', 'Backup `{0}` has been deleted', $file['filename']));
             }
 
             $io->success(__d('database_backup', 'Deleted backup files: {0}', count($files)));