+++ src/Driver/Sqlite.php
@@ -34,12 +34,14 @@ class Sqlite extends AbstractDriver
 
         //Drops each table
         foreach ($schemaCollection->listTables() as $table) {
-            array_map([$connection, 'execute'], $schemaCollection->describe($table)->dropSql($connection));
+            /** @var \Cake\Database\Schema\TableSchema $TableSchema */
+            $TableSchema = $schemaCollection->describe($table);
+            array_map([$connection, 'execute'], $TableSchema->dropSql($connection));
         }
 
         //Needs disconnect and re-connect because the database schema has changed
-        $connection->disconnect();
-        $connection->connect();
+        $connection->getDriver()->disconnect();
+        $connection->getDriver()->connect();
 
         return true;
     }